refactor: dedupe model-request OTel span between `Instrumentation` and `InstrumentedModel`
Both `Instrumentation.wrap_model_request` (agent flow, capability-driven) and
`InstrumentedModel.request`/`request_stream` (standalone / `direct.model_request*`)
were independently building the same `chat <model>` CLIENT span — same attributes,
same `finish(response)` closure, same cost/metrics ordering. With #5427 making
`InstrumentedModel` permanent, this duplication is no longer transitional.

Extracts the shared logic into `_instrumentation.open_model_request_span` (a
`@contextmanager` that yields a `finish(response)` callback). Both call sites
now do one `prepare_request`, open the helper, run the wrapped request, call
`annotate_tool_call_otel_metadata` + `finish`.

Drops `InstrumentedModel._instrument`. Trims a chunk of now-unused imports
in both files.
