+++ pydantic_ai_slim/pydantic_ai/capabilities/instrumentation.py
@@ -2,29 +2,22 @@
 
 from __future__ import annotations
 
-import warnings
 from collections.abc import Awaitable, Callable
 from dataclasses import dataclass, field, replace
-from typing import TYPE_CHECKING, Any, cast
+from typing import TYPE_CHECKING, Any
 
 from opentelemetry.baggage import set_baggage as _otel_set_baggage
 from opentelemetry.context import attach as _otel_attach, detach as _otel_detach
-from opentelemetry.trace import SpanKind, StatusCode
+from opentelemetry.trace import StatusCode
 from pydantic_core import to_json
 
 from pydantic_ai._instrumentation import (
-    GEN_AI_REQUEST_MODEL_ATTRIBUTE,
-    GEN_AI_SYSTEM_ATTRIBUTE,
-    MODEL_SETTING_ATTRIBUTES,
-    CostCalculationFailedWarning,
     InstrumentationNames,
     annotate_tool_call_otel_metadata,
-    build_tool_definitions,
     event_to_dict,
     get_agent_run_baggage_attributes,
     get_instructions,
-    model_attributes,
-    model_request_parameters_attributes,
+    open_model_request_span,
     serialize_any,
 )
 from pydantic_ai.exceptions import ApprovalRequired, CallDeferred, ToolRetryError
@@ -259,7 +252,6 @@ class Instrumentation(AbstractCapability[Any]):
         request_context: ModelRequestContext,
         handler: WrapModelRequestHandler,
     ) -> ModelResponse:
-        settings = self.settings
         model = request_context.model
 
         # Track the latest messages so _run_span_end_attributes has them on error paths
@@ -276,99 +268,13 @@ class Instrumentation(AbstractCapability[Any]):
         # instead of falling back to reading `ModelRequest.instructions` from history.
         self._last_model_request_parameters = prepared_parameters
 
-        operation = 'chat'
-        span_name = f'{operation} {model.model_name}'
-        attributes: dict[str, Any] = {
-            'gen_ai.operation.name': operation,
-            **model_attributes(model),
-            **model_request_parameters_attributes(prepared_parameters),
-            **get_agent_run_baggage_attributes(),
-            'logfire.json_schema': to_json(
-                {
-                    'type': 'object',
-                    'properties': {'model_request_parameters': {'type': 'object'}},
-                }
-            ).decode(),
-        }
-
-        tool_definitions = build_tool_definitions(prepared_parameters)
-        if tool_definitions:
-            attributes['gen_ai.tool.definitions'] = to_json(tool_definitions).decode()
-
-        if prepared_settings:
-            for key in MODEL_SETTING_ATTRIBUTES:
-                if isinstance(value := prepared_settings.get(key), float | int):
-                    attributes[f'gen_ai.request.{key}'] = value
-
-        record_metrics: Callable[[], None] | None = None
-        try:
-            with settings.tracer.start_as_current_span(span_name, attributes=attributes, kind=SpanKind.CLIENT) as span:
-                # `finish` is a closure rather than inline so we can (a) set result attributes
-                # inside the `with span:` block — they attach to the span — and (b) call the
-                # captured `record_metrics` in the outer `finally` AFTER the span closes,
-                # so observability backends that aggregate metrics from span attributes
-                # don't double-count.
-                def finish(response: ModelResponse) -> None:
-                    nonlocal record_metrics
-
-                    # FallbackModel updates these span attributes via get_current_span().
-                    attributes.update(getattr(span, 'attributes', {}))
-                    request_model = attributes[GEN_AI_REQUEST_MODEL_ATTRIBUTE]
-                    system = cast(str, attributes[GEN_AI_SYSTEM_ATTRIBUTE])
-
-                    response_model = response.model_name or request_model
-                    price_calculation = None
-
-                    def _record_metrics() -> None:
-                        metric_attributes = {
-                            'gen_ai.provider.name': system,
-                            'gen_ai.system': system,
-                            'gen_ai.operation.name': operation,
-                            'gen_ai.request.model': request_model,
-                            'gen_ai.response.model': response_model,
-                        }
-                        settings.record_metrics(response, price_calculation, metric_attributes)
-
-                    record_metrics = _record_metrics
-
-                    # Compute cost before the `is_recording()` gate so `_record_metrics`
-                    # always emits cost data, even when the span is dropped by sampling.
-                    try:
-                        price_calculation = response.cost()
-                    except LookupError:
-                        pass
-                    except Exception as e:  # pragma: no cover — safety net for unexpected genai_prices errors
-                        warnings.warn(
-                            f'Failed to get cost from response: {type(e).__name__}: {e}',
-                            CostCalculationFailedWarning,
-                        )
-
-                    if not span.is_recording():
-                        return
-
-                    settings.handle_messages(request_context.messages, response, system, span, prepared_parameters)
-
-                    attributes_to_set: dict[str, Any] = {
-                        **response.usage.opentelemetry_attributes(),
-                        'gen_ai.response.model': response_model,
-                    }
-                    if price_calculation is not None:
-                        attributes_to_set['operation.cost'] = float(price_calculation.total_price)
-
-                    if response.provider_response_id is not None:
-                        attributes_to_set['gen_ai.response.id'] = response.provider_response_id
-                    if response.finish_reason is not None:
-                        attributes_to_set['gen_ai.response.finish_reasons'] = [response.finish_reason]
-                    span.set_attributes(attributes_to_set)
-                    span.update_name(f'{operation} {request_model}')
-
-                response = await handler(request_context)
-                annotate_tool_call_otel_metadata(response, prepared_parameters)
-                finish(response)
-                return response
-        finally:
-            if record_metrics:
-                record_metrics()
+        with open_model_request_span(
+            self.settings, model, request_context.messages, prepared_settings, prepared_parameters
+        ) as finish:
+            response = await handler(request_context)
+            annotate_tool_call_otel_metadata(response, prepared_parameters)
+            finish(response)
+            return response
 
     # ------------------------------------------------------------------
     # wrap_tool_execute — tool execution span
